+++ Prototype/src/System/graphics/graphicsEngine.cpp
@@ -12,10 +12,10 @@ using namespace::std;
 vector<set<Renderer*>> GraphicsEngine::renderQueue = vector<set<Renderer*>>();
 
 //Screen dimension constants
-const int GraphicsEngine::SCALE_X = 3;
-const int GraphicsEngine::SCALE_Y = 3;
-const int GraphicsEngine::SCREEN_WIDTH = GraphicsEngine::SCALE_X * 512;
-const int GraphicsEngine::SCREEN_HEIGHT = GraphicsEngine::SCALE_Y * 288;
+const float GraphicsEngine::SCALE_X = 3;
+const float GraphicsEngine::SCALE_Y = 3;
+const int GraphicsEngine::SCREEN_WIDTH = (int)(GraphicsEngine::SCALE_X * 512);
+const int GraphicsEngine::SCREEN_HEIGHT = (int)(GraphicsEngine::SCALE_Y * 288);
 //Frame rate related
 const int GraphicsEngine::SCREEN_FPS = 60;
 const int GraphicsEngine::MICROSECONDS_PER_FRAME = 1000000 / SCREEN_FPS;
