+++ Prototype/src/System/graphics/graphicsEngine.h
@@ -23,8 +23,8 @@ public:
     inline static LWindow* getActiveWindow() { return activeWindow; }
 
     //Screen dimension constants
-    static const int SCALE_X;
-    static const int SCALE_Y;
+    static const float SCALE_X;
+    static const float SCALE_Y;
     static const int SCREEN_WIDTH;
     static const int SCREEN_HEIGHT;
